Added renderer, engine and editor folders. Added application and renderer class. Adjusted CMakeLists.
