+++ editor/src/application.h
@@ -1,0 +1,14 @@
+//
+// Created by jeang on 2024-01-25.
+//
+
+#ifndef GRAPHICS_PLAYGROUND_APPLICATION_H
+#define GRAPHICS_PLAYGROUND_APPLICATION_H
+
+
+class application
+{
+};
+
+
+#endif//GRAPHICS_PLAYGROUND_APPLICATION_H
