+++ engine/CMakeLists.txt
@@ -1,0 +1,2 @@
+cmake_minimum_required(VERSION 3.26)
+project(engine VERSION 0.1)
