+++ projects/CMakeLists.txt
@@ -1,0 +1,2 @@
+cmake_minimum_required(VERSION 3.26)
+project(projects VERSION 0.1)
