+++ renderer/CMakeLists.txt
@@ -1,0 +1,9 @@
+cmake_minimum_required(VERSION 3.26)
+project(renderer VERSION 0.1)
+
+set(CMAKE_CXX_STANDARD 23)
+
+add_library(${PROJECT_NAME} SHARED src/renderer.cpp)
+
+target_include_directories(${PROJECT_NAME} PUBLIC include)
+target_include_directories(${PROJECT_NAME} PRIVATE src)
