+++ renderer/include/renderer.h
@@ -1,0 +1,14 @@
+//
+// Created by jeang on 2024-01-25.
+//
+
+#ifndef GRAPHICS_PLAYGROUND_RENDERER_H
+#define GRAPHICS_PLAYGROUND_RENDERER_H
+
+
+class renderer
+{
+};
+
+
+#endif//GRAPHICS_PLAYGROUND_RENDERER_H
